Refactoring. Moving Route class to RouteSegments.

+++ routing/routing_integration_tests/osrm_route_test.cpp
@@ -302,7 +302,7 @@ namespace
     Route const & route = *routeResult.first;
     IRouter::ResultCode const result = routeResult.second;
     TEST_EQUAL(result, IRouter::NoError, ());
-    TEST_LESS(route.GetTotalTimeSec(), numeric_limits<uint32_t>::max() / 2, ());
+    TEST_LESS(route.GetTotalTimeSec(), numeric_limits<double >::max() / 2.0, ());
   }
 
   // There are road ids in osrm which don't have appropriate features ids in mwm.
